Fix some regressions from the last merge from upstream
Some of the changes I had to make were obvious (like trying to assign
a std::string to a char*), some other changes were... the result of
a very unproductive debugging session trying to fix some random data
strings that probably wouldn't have happened if this code wasn't
rewritten to do things as much the C way as possible. These changes
from `const char*` to `std::string` are basically necessary because
.c_str() apparently has this pitfall where you can't keep it around at
all and would need to strcpy it, so why bother with that
`static const char raw[]` assignment as well anyway.