Remove duplication of wordwrapping algorithms and PrintWrap
Basically, Misa needed a wordwrapping algorithm at some point, and
implemented one from scratch, but only needing to have PrintWrap in
mind. So, it iterates over the lines and copies lines to a buffer one
by one. However, this may not work for all places where I was already
using the existing (in the localization branch) wordwrapping function
to give a string with newlines inserted. So now there are still two
functions to do wordwrapping, but mine uses the new one
(Graphics::next_wrap) as a backend. I also made next_wrap take
loc::langmeta.autowordwrap into account, by simply not treating spaces
as anything special if it's enabled.

I also renamed Graphics::wordwrap, Graphics::wordwrap_balanced and
Graphics::unwordwrap to be prefixed with string_. next_wrap and
next_wrap_s should probably also be renamed next_wordwrap and
next_wordwrap_s to make their purpose a bit more clear, but that's
probably out of scope of this branch (it's gonna be large enough as is)

(Personally, I think the next_wrap/next_wrap_s functions are a LOT
harder to understand, with all the pointers-across-functions juggling
that it does, and it has not been documented to alleviate that at all.
But no need for me to touch it further now, since the main problem of
having to be stuck with two separate algorithms has been resolved.)