Reorganize location of language files directory
My original idea was to have the language files inside data.zip, and,
like any other asset, allow translators or developers to override them
by placing them in their regular assets folder (like Documents/VVVVVV)
instead. However, this is very much not ideal:

- Developers wanting to update the language files (especially when they
  want to sync them from the in-game maintenance menu) need to first
  copy the files from the repo into their assets folder (out of version
  control), update the files, and then copy them back in so their
  changes can actually be in the repo. Or, do what I did, run a
  shellscript to hardlink each language file between my assets folder
  and the repo...
- Translators also need to work outside data.zip of course, and they
  would also need to copy everything to their assets folder first.
- data.zip would need to be updated for every single text or
  translation update.
- The lang folder in the repo is just sitting there. It's never "used",
  it always requires copying out of version control, because there's
  nowhere to place them in the repo where the game will read it.

So the better idea is this: VVVVVV should be distributed with the lang
folder in the same place where data.zip is (so for example, in the
user's Steam folder), and the game will use that if it exists.

If it doesn't, that probably means someone has compiled VVVVVV and
created a build folder for themselves. In that case, the game checks
if /desktop_version/ (or \desktop_version\ on Windows) is somewhere up
in the path, and if it is (which is likely, since the build folder was
probably made inside of desktop_version/), the game will try to use
desktop_version/lang/. Which is convenient, because that's the
up-to-date version that belongs to the code you compiled, and it's
under version control!

(And I'm glad desktop_version/ exists, since you cannot assume that the
root folder is named "VVVVVV". I saw "the letter v six times i guess"
the other day. The alternative would be to just check the parent
folder, but I think VVVVVV.exe goes a bit deeper than one level at
least on Windows. And traversing all the way up to the filesystem root
sounds like a bad idea.)

It's also possible to override the location of the lang folder by
passing a -langdir argument to VVVVVV.

If the game finds the lang folder inside desktop_version/ instead of
next to data.zip, it will also unlock the translation maintenance menu.
There's now also the argument -translator, which does the same thing,
and it doesn't need translators to extract/move any files, because with
lang/ being outside data.zip, the game can also write to it. But right
now, it can't do that yet, so I'll still need to change that.