Fix language sync option for new lang folder system
This also adds writeDir to FileSystemUtils.cpp, which is basically
what it says: the base directory that VVVVVV can write stuff to. Except
when syncing language files, that's when we temporarily switch it out
for the main language dir. The main PHYSFS page says this: "While you
CAN change the write dir as many times as you like, you should probably
set it once and stick to it." We have good reasons to change it here,
plus it's only in a debug menu pretty much.

Also, yes, this commit time is technically before the last one.