Add menu test mode
Activating this mode will let you cycle through most of the menus, to
be able to spot problems with translations in them. This is especially
important for the menu system, since the width limit for each option is
hard to define, and options that stick out from the rest may look bad
even when they are within limits.