Start migrating localization system from std::map to C-HashMap
I decided it would be better to use something else than `std::map` for
storing translations, and ways to do without it (that weren't naïve or
convoluted) started to take more and more shape, and then I ended up at
this library: https://github.com/Mashpoe/c-hashmap, which worked very
well with what I had in mind.

The idea is that we store all strings somewhere managed, and then the
hashmap only needs pointers to those strings. For storing strings, I
created a `Textbook` structure, which consists of one or more 50 KB
"pages" (allocated as needed) on which you can simply write strings in
both languages back-to-back with `textbook_store(textbook, text)` and
get pointers to each of them. (I was originally going to just use one
big buffer and realloc to double the size when filled up, but then the
hashmap would be full of dangling pointers...) When needed, like when
switching to a different language, an entire textbook can be freed at
once. There's one main textbook right now, and we can add another
shorter-lived one (and corresponding hashmaps) for text in custom
levels.

Textbooks are also useful for storing room names, so the room names
array is now an array of char pointers instead of 161-character
buffers.

The only std::map that I didn't migrate yet is the cutscenes map, but
that probably wouldn't be too hard anyway.