Add typing support to room name translation mode
In edit mode, you can now press Enter or E to edit the room name.
It just doesn't save it to a file yet, but it does associate the new
name with that room until the language files are reloaded!

If the current room's name is untranslated, it'll show up in cyan.