Freeze physics in edit mode (room name translation mode)
By not pausing the game, some rooms would become practically impossible
to translate (like vertical tunnels or rooms with otherwise no place to
stand, even on spikes, like "Exhausted?"). This was a bit more tricky
to solve because the map and pause menu use a different gamestate in
order to pause the game, but I wanted the game to remain in GAMEMODE.

To allow for similar problems to be easier to solve in the future (and
to avoid tangling room name translation mode into the code), I added a
Game::physics_frozen() which returns whether everything should be
frozen. This makes gamelogic() stop, and avoids >30fps interpolation
still being visible by changing Graphics::lerp. (I had to ask Misa how
to solve that one)