Add full saving support to room name translation mode, add "expl mode"
Room name translation mode is now mostly functional! Translating a room
name will save it to the language file, the explanations are now loaded
from file (instead of displaying a placeholder), and I added something
new as well: explanation mode! This mode can be used to edit the
explanations given for each room name at the top of the screen, and
editing an explanation will save it to ALL languages. You can switch
between naming mode and explanation mode with Ctrl+E.