Add room name translation/explanation counters
The "XXX left" at the top now works, and reflects the number of
untranslated or unexplained room names, depending on name/expl mode.
It's also possible to mark a room as not needing an explanation with
a single period, so the room will not have an explanation but the
counter of unexplained rooms will still decrease.