Add Ctrl+X shortcut to text input
So, this kinda doesn't have to do with the localization system (since
it affects text input throughout VVVVVV) but since it might be a pain
to have to backspace a long room name explanation, I added it. Ctrl+X
now cuts the entire text input to the clipboard, so you can paste it in
a real text field somewhere, edit it, and paste it back in with the
Ctrl+V shortcut that already existed.