Fix off-by-one error when clearing room name array
I was wondering why the untranslated counter was a few rooms off, that
was probably why, I didn't clear the entire array so it had some
uninitialized memory in it. Actually, those were uninitialized pointers
that would later be indexed because they weren't NULL. That totally
should have segfaulted on me.