Move and update languages README
I moved it out of the English folder and into the main languages
folder, so it doesn't get copied two dozen times, and I did a major
update to this file to make it really a good documentation of how to
make and maintain translations. Some of the information had already
become outdated as design decisions changed (the PR description needs
an update too).

The past weeks I've also been thinking a lot about fonts - for example,
I really don't want to require data.zip to be continuously updated for
languages to work properly now, that would just keep causing issues and
desynced versions with missing fonts - so things are a little more
complicated: we want to use the Unicode font by default, but if it's
missing then we should fall back to the font in data.zip, but a level-
specific graphics/font.png should *still* get priority over the new
font even when we don't want it from data.zip, and it should also not
make Russian menus unreadable (so: two 8x8 fonts may be loaded). And
how CJK fonts should be defined, where should the game find them, what
if we want to support non-monospaced fonts, all while 2.3 already has
a new system for Unicode fonts. But I think I have all that pretty much
fully thought out now, so that'll save time during implementation
later!