Only accept directories as languages in the language folder
The game needs to ignore README.txt now, and language packs can only
be folders.

I just realized Apple's .DS_Store files exist, so this helps with that
too, but there was also something with hidden duplicates of folders now
that I think about it, so I still need to look into that.