Make loc::gettext take and return a const char* instead of std::string
The translations are now stored as raw C strings, so we can save
ourselves a lot of conversions (which even manifests itself as code
simplification in many places by being able to remove .c_str()).
Whenever a std::string is needed, it's either implicitly casted, or, in
a few cases, I've inserted a std::string constructor locally, which is
basically the same as what was already inside loc::gettext until now.