Add plural forms system
Initially I wanted to implement it with per-language formulas like
ngettext does, hardcode the 19 or so possible formulas covering 160
languages, and just let each translator choose one of them. But
recently I realized that for understandability, flexibility, and less
hardcodedness (I am also not adding an expression parser to VVVVVV lol)
we can just let translators manually assign plural forms to the numbers
that they have to translate anyway. This is probably not so cumbersome,
you can just do a big find-and-replace on the form that's most common,
and it will very likely also be possible to just change cells in Excel.

I don't think we need any higher numbers than 100 anyway (we don't have
possibilities for "121 unread messages" or something), and if we go
beyond 100 then most of the time we already use "Lots", which can now
also get its own form if needed. It'd even be possible to assign
different forms for longer numbers so the common 0-20 can use a normal
string for stuff like "Twenty crewmates remain" but the uncommon 21-100
can use "Seventy Eight crewm. left" or something like that.

Also, I improved the time trial results screen a bit, making it less
tight for translations and allowing the rank to be more than one letter
without going off-center.