Simplify XML structure in language files
The main change is that translations are no longer in between opening
and closing tags, but as an attribute next to English. This has a
couple of advantages:

- It's more consistent and easier to understand
- The translation is right next to the English string, instead of being
  separated by the other arguments like the explanation
- I (and maybe others) don't have to be annoyed anymore by XML parsers
  automatically deciding to self-close tags that were supposed to be
  left open, whenever programmatically editing the file
- Makes it easier to let the language files be imported and exported by
  Excel. This was my original inspiration for doing this.

So, I've been experimenting with Excel, I knew you could work with XML
via macros, but recently I found out it also has a feature to add XML
schemas (XSD) so you can import and export XML files directly without
having to change your XML structure. Sounds great, except XSD is
convoluted, and it's hard to get Excel to do what you want. Having only
attributes, or only tag content (or child tags), is doable, and exactly
what Excel was designed to let you do with relative ease. Having both
attributes and tag content requires special care to get right in XSD,
and once you do get it working, it will result in a weird column
ordering, the translation header will be the tag's name, exporting as
XML also doesn't work most of the time...

Another simplification I did was to remove comment attributes. It would
require special code in the language sync feature to load them in and
map them out so they can be kept when syncing, which means more
complexity, so I just decided to scrap it in favor of manually keeping
notes. It's also a question whether translator's comments should be kept
around in the end product and if not, how to remove them (an option in
the menu?) and when (by whom)... Plus, translators may always use CAT
tools anyway.