Add textbox padding function
This will make it easier to add support for localizing some pretty
annoying textboxes. Instead of writing:

    graphics.addline("     You have rescued  ");
    graphics.addline("      a crew member!   ");

You now write:

    graphics.addline("You have rescued");
    graphics.addline(" a crew member! ");
    graphics.textboxpad(5, 2);

This is now done for the Level Complete screen as a proof-of-concept.
Another function I've been thinking of adding is one to center the text
in a textbox by padding with spaces as well. In that case you'd get
something like:

    graphics.addline("You have rescued");
    graphics.addline("a crew member!");
	graphics.textboxcentertext();
    graphics.textboxpad(5, 2);

And then we're just left with the regular textbox lines that need to be
looked up and rewrapped, just like already happens with dialogues in
scripts.