Add textbox padtowidth and centertext functions
These _should_ be enough to reproduce every manually padded or centered
textbox in the game in automatically wordwrapped translations. And they
keep larger fonts in mind for CJK (hence the pad-to-pixels, and not
pad-to-characters - we don't want to specify that a textbox needs to be
35 characters wide when only 26 will fit onscreen in Chinese or
Japanese - while padding by characters makes sense since it's already
based on the text itself).

So, padtowidth(size_t new_w) takes a width in pixels, and tries to pad
each existing line of text in the textbox so it becomes that width.
This is mainly intended for the terminal headers like these:

"        -= PERSONAL LOG =-         "
"      -= Research Notes =-       "
"     -= Personal Log =-      "
"          -= WARNING =-         "
"   -= D.S.S. SOULEYE =-   "

centertext() centers the existing text based on the current width of
the textbox.

My plan is to let these functions also be controlled by translators in
cutscenes.xml. For example:

<dialogue speaker="gray"
	english="-= Research Notes =-"
	translation="-= Notoj de esplorado =-"
	padtowidth="264"/>

<dialogue speaker="gray"
	english="...everything collapses, eventually. [...]"
	translation="...ĉio fine kolapsos. [...]"
	centertext="1"
	pad="1"/>

The first example would be padded to 264 pixels (33 characters), and
the second example would be automatically wordwrapped (with 2
characters less space than usual), then centered in place, and then
padded on both sides.