Improve balanced wordwrapping algorithm
So, the original idea was to define minimum and maximum bounds and find
the wrapping limit where the text is written to the end of each line as
much as possible. Probably a much simpler and better-working idea: have
a maximum limit only, see how many lines are needed for that limit, and
then just make the limit smaller until you start needing an extra line.