Auto-unwordwrap English dialogue text in XML, fix unwrapping algorithm
This allows us to write stuff like this as the key:

"-= SHIP RADIO =-

[ Status ]
Broadcasting"

Every English dialogue from the scripts gets unwordwrapped before being
looked up in the translation maps, so we don't have hard wrapped text
in the language files. But the unwordwrapping algorithm can't determine
whether the bottom two lines are just a hard wrapped sentence or
something that's supposed to be on two different lines. But we still
want to present translators a non-mangled version of the string (one
that doesn't have "[ Status ] Broadcasting" in it). By auto-unwrapping
the strings in the language files as well when loading them in, this is
now possible.

Originally I thought I was going to be adding an internal scripting
command or two, and editing some main game scripts. But now it looks
like I'm not even gonna have to touch the main game scripts at ALL,
which wasn't even a requirement (I'd have been perfectly fine with
adding some extra commands for certain special cases for example)

The unwordwrapping algorithm also had an issue where it could leave a
space after two newlines. This is now also fixed.