Add support for cutscene dialogue disambiguation
Ladies and gentlemen, I've successfully left my mark on VVVVVV's code.

I've added an internal scripting command.

Right after the commit where I said "I might not even have to add any
commands nor touch the scripts themselves for the localization branch!"
I (re-)discover that the [C[C[C[C[Captain!] dialogues are misaligned if
they're not padded with spaces on the right. That wouldn't be a
problem, I added support for pad_left and pad_right attributes for
cutscene translations despite thinking they weren't going to be used in
scripts. But there was a design compromise for simplicity: if the same
text appears multiple times in the same cutscene, it had to be
translated the same. And then I checked the Nicalis translation and
discovered that in ALL four cases where the same text is used multiple
times in scripts, all the translations were made unique. Even COMBINE.

It kinda makes sense that text might need to change because of cultural
reasons, or simply because it's that much more natural to not repeat
the same word in another language... And translators probably won't
expect their first translation of an identical line to be swallowed by
the second. So now there's a textcase(x) scripting command which can be
placed above the text command to specify a specific occurrence of the
dialogue, and the XML can now have the attribute case="x".

Internally, this uses the same principle as the plural system: the
English key is prefixed by a one-byte disambiguator. By default all
dialogues are case 1 (it's not a null terminator). There's now an
add_disambiguator(...) which is used for plural forms and cutscenes.