Add failsafe when setting lang write dir if VVVVVV/lang exists
The localization branch no longer uses the VVVVVV folder as a place for
translators to copy their language files, but Reese thought it still
did. That overrules the repo's language folder when loading, which
causes room name translations to be continuously and silently
overwritten with empty ones because the saved ones are never reloaded.

Now at least you get a general error message, and a specific one in the
console, which should do - you don't continue translating room names,
because you know something is going wrong and will have to be checked.