Add loadtext(lang) and iflang(lang,script) scripting commands
Just displaying a custom level in VVVVVV's language without any control
possible may be just a bit too simple for some levels - you may want
to decide the priority, translate to a language that VVVVVV doesn't
support, or do something else. So to give level creators some
flexibility, I've decided to add two commands:

- loadtext(lang): load a translation for the level by language code.
  Pass an empty language code to revert to the default behavior of
  simply using VVVVVV's language.

- iflang(lang,script): if VVVVVV's language is set to lang, go to a
  custom script. This is not affected by the language code you pass to
  loadtext(), only by what language the user has selected in the menu.

These commands are both internal and simplified and work the same way
as both. If I'm not mistaken, these are also the first simplified
commands that have been added since 2.1. Forwards compatibility is
well-preserved unless you're choosing to break it as a level creator:
when playing a translated level in an older VVVVVV version, it will
just logically be fully in English (or the original language of the
level), and iflang() "behaves" as if the language is set to something
else unforeseen by the level creator (which you should probably account
for anyway, since new languages can be added in new VVVVVV versions).

That's probably enough for custom level localization in this PR -
there's plenty to play with, and what isn't yet translatable isn't
immensely important.