Allow "Level Complete!" and "Game Complete!" to be translated
These were a bit harder since they've always used an image with the
text instead of simply bigprint'ing it (I guess to avoid having to add
a special function to print bigger text, which ended up being added
anyway)

The images are still used in English (or if the translation happens to
be the exact same), and custom level assets still override the
translation, but otherwise, the translation is simply printed.

Since these are very prominent strings with a kinda tight limit of 18
characters, instead of simply letting them go outside the box, they can
be displayed in a smaller font size, just in case a translator misses
the limit and it's not noticed/checked before somehow ending up in the
final version. Or, in case a translator really can't make it fit within
the limit.