Improve plural forms system for numbers above 100
I originally designed the plural forms system with the assumption/
simplification that there were no strings in VVVVVV where plural forms
mattered and that could also reasonably go higher than 100 instead of
stopping at Lots. But "Hardest Room (with %d deaths)" is such a string.
It was hardcoded to be plural in English since you can kinda get away
with "1 deaths", but in languages with 3 or more plural forms that also
distinguish between 121 deaths and 122 deaths, it's probably a problem.

Fortunately, the system doesn't need a massive overhaul. In all
languages supported by Pootle (160), the plural rules always repeat
themselves every 100 numbers. So we can just copy 0-99...

Nah, it's not "101 trinket". The numbers 000-019 can be exceptions to
this repeating pattern, so translators now only have to provide forms
for the numbers 101-119 to complete the circle. You can provide forms
up to 199, but otherwise they're just copied from 0-99.