Make time trials menu translatable
This involved both a refactoring of the code and a redesign of how the
results are displayed. I first deduplicated all the results-displaying
code (it was copy-pasted for all time trials and had some hardcoded
strings for par times and trinket counts.) Then I moved things around
to make it less tight and also more dynamic for translators (it may
still be tight for some languages, but it should be much better than
the old screen.)

Main visual differences are that the "BEST" is removed from all the
best individual stats and replaced by a header "RECORDS", and the par
time and best rank are now columns of their own. The numbers now move
depending on the lengths of the labels to the left of them. Also, it
may be a nice change that the par time is now closer to your best time.