Make PrintWrap give next line y position as return value
This allows you to put a specific amount of spacing after a wrapped
text regardless of how many lines that text is. This is now also done
in the menus, where all the lines like "Currently ENABLED" used a
hardcoded position, which could look ugly in other languages because
there's randomly no spacing, one or two lines of spacing, and the sub-
line may also randomly move based on the English version and not the
translated one.

Also, I changed the PPPPPP and MMMMMM display to not use
VVV_fillstring. Partially because it went too far in whatever it was
trying to optimize (bytes of executable size? avoiding repetition of
"Current soundtrack:"?). And also because we don't know if the
translations should be something like "MMMMMM-ov" and "PPPPPP-ev" in
another language.