Change PrintWrap default arguments for linespacing and maxwidth to -1
Like I said in the previous commit, I don't want to see every PrintWrap
call repeating the same hardcoded `10, 304` when someone decides that
default function arguments are bad - so this encourages a compromise:
if it does eventually happen, at least repeat `-1, -1` instead, which
indicates that the function should choose the linespacing and maxwidth
automatically. I think PrintWrap is basically the best example of when
to use default arguments: when the function is called in many places,
some of the arguments are almost always the same, they're some weird
values (other than 0/-1/NULL/etc) that might change in the future, and
you want to avoid repeating those same values ad nauseam. Switching to
-1 also prepares for CJK changes because the implementation of
PrintWrap will probably need to choose a higher value than 10 for
different fonts.