Make language screen show on first startup
This will obviously be necessary, since people might not be able to
read the menus to find language selector. I've seen an astonishingly
high number of games/software/devices just default to English while
having a language option somewhere buried in the English menu.

In addition to <lang>, there's now also a <lang_set> in the settings
file, which stores that the language has alreay been set once.

After the initial language screen, it jumps to the title screen that
explains what ACTION is in the chosen language. So the loading screen
is skipped that time.