Overhaul language screen for two columns
There are almost definitely going to be more languages than would fit
on the screen in one column, and even if not, there needs to be room
for people to add their own translations (like pirate/lolcat/etc, or
just otherwise unofficial ones). And demoting some languages to a
"second page" is probably very undesirable. So the language screen now
has two columns, and 4-directional movement. This obviously needed some
changes in the menu system, but it's not all that extreme; menu options
just need to be placed at different coordinates than usual in one menu,
and there needs to be a special case in input handling.

There are also a bunch of dummy language options for now, so the menu
can be tested. I'll soon hopefully also be able to test the movement
with a Wingman joystick on a Windows XP cardboard box, just in case.