Make rolling credits translatable
"Thanks for playing!" is now split on two lines to make it not so
tight (even in English), and the max position is now fixed (it was
already off by 16 pixels and needed to account for the extra line now
too). This is now another of the rare cases where translators need to
translate two halves of a string separately, but it's clearly
indicated.

I've also made the "big names" on the first few pages of the regular
credits UNtranslatable. The idea was that these names finish a
sentence and might therefore need grammar rules applied to them, or
translators might want to transliterate them. And why not, there's only
a few big names. But there are also a lot of patrons and GitHub
contributors, which would not be translatable unless we added every
single one of them to the language files. You could see those names as
something different because they're a big list... But I checked Dicey
Dungeons and only the "special thanks" on the last page has some
English in it that is correspondingly translated, none of the regular
names have been changed in any language as far as I could find. So it's
not really necessary (unless a translator says it is) and otherwise
there's probably a way to reword.