Remove extraction test mode
I like the principle of the test mode, but I find it being in the way
of making code simplifications and abstractions, it adds complexity
without even having complete coverage, and it's not really THAT useful
over just doing a manual check whenever something looks untranslatable.
So I just decided to scrap it and allow the code to be less complex.
Maybe it was just too early for this feature to exist and it can be
re-added later, but it probably won't be necessary.