Add language file reload button F12, work on strings.xml
If and only if the translator menu is available, you can now press F12
anywhere in the game to reload the language files. This is convenient
for translators, since they can immediately preview most of their
changes by just saving the language files and then pressing F12.

I also added more explanations and clarifications to strings.xml, and
worked on the Dutch translation.