Make more strings translatable that weren't yet translatable
This includes trophy titles and descriptions; one change I made was to
split the trophy descriptions into title at the top, and description at
the bottom. They were both at the top before, but there's probably not
enough room for that in many languages (some of the English
descriptions are already close to the limit). I've made some other
small changes as well, like including the % sign on the loading screen
in the translation (I think some languages actually want to display
"100 %" instead of "100%")

I've also been working on the Dutch translation in the meantime, many
cutscenes are now translated as well (not really proofread yet but see
this as a draft)