Make language metadata apply on the language screen
One little oversight at first was that there are some language metadata
options that affect the language screen too - they should be the
metadata of the current menu option, not the metadata of the already
configured language. For example, choosing a language with automatic
uppercasing disabled shouldn't make all the language options lowercase,
and having wordwrapping disabled shouldn't affect languages that assume
their credits to be wordwrapped. This is now easily solved with
`LangMeta* loc::get_langmeta(void)` that usually just returns a pointer
to `loc::langmeta`, but on the language screen just gives the metadata
to the selected language. All places that used `loc::langmeta` now use
this getter instead.