Add text limits check to translator menu
This feature allows you to check the entirity of strings.xml and
strings_plural.xml for text that doesn't fit within the bounds that
many of the strings have. And after I implemented that, I decided it
was a good idea to add a "global" version of this feature in the
maintenance menu as well (which doesn't limit itself to one language).

This is a feature that a lot of software and games don't have, but it
was a big one on my list, and I didn't want to cut corners here by not
adding it - it may catch some text overflows that would otherwise not
be caught before 2.4 and make translations seem sloppy, *especially*
where written out numbers are added to strings (like "You collected
Seventeen trinkets" running offscreen). And VVVVVV doesn't have the
highest resolution (it can never fit more characters on a line than
Teletext)

On a technical level, VVVVVV can now parse the `max` attributes that
were already on strings and check if the translation (or in the case
of strings_plural, the translations for the needed numbers, less than
or equal to the newly added attribute `expect`) fits within the bounds.
If not, it's added to loc::text_overflows. This check is only done when
the limits check feature is actually used.