Use Unicode accented characters in credits
This replaces the instances of {|}~ by öäÁå, and changes
"Remi Verschelde" into "Rémi Verschelde".

This change requires that the traditional font is superseded across the
board, which is exactly what I'm planning to do - basically the only
real item left on my todo list for this PR (apart from some pretty
small ones) is to add Space Station via roughly the same method as how
the language files are included. So might as well do it now.

Space Station also has accented character glyphs for {|}~ to keep
compatibility with existing levels that use them, so I'd only have to
do the é - but we may not make the same changes to the CJK fonts (why
would we? Existing levels would just default to using the 8x8 font),
and I'd rather consistently use the proper characters instead of
practically using our own mini lookup table for certain ones.