Add fonts directory to repo
This adds https://github.com/Fussmatte/space-station-font so VVVVVV can
now always display Unicode characters without font customizations!

On a technical level, this works the same as the lang directory, which
is explained in f4b423032bac6f1251107ca12f8a3183601e5959.

In officially distributed versions, this fonts folder should be present
next to data.zip. In self-compiled versions, the game will try to find
it in /desktop_version/, and thus automatically use the most recent
version, without having to download anything extra or update data.zip.

This fonts folder is mounted as "graphics/" after the user's VVVVVV
directory but before data.zip. That is logical, because we are just
merging different assets locations into one. Physically (on disk) the
fonts are in another folder so they're neatly categorized, but that's
all merged into one graphics folder that the game can work with.

What it also does is solve a lot of potential problems with new assets:
- We avoid having to keep updating data.zip for every font update we
  may need to do (adding a language for example) and having all sorts
  of different data.zip versions floating around while having to keep
  telling people to download the latest version.
- It purely uses the layered moddable filesystem to handle priority,
  instead of dedicated code to prioritize level-specific
  graphics/font.png over VVVVVV's stock fonts/font.png, or
  level-specific fonts/ over level-specific graphics/, for example.
- Again, fonts are subject to change when new languages are added or
  when a CJK font is missing a character that translators want to use
  or something like that, so they better somehow automatically update
  right in sync with the code, no matter what happens or what a user
  without custom fonts does or fails to do (or re-download) when they
  pull changes to their self-compiled versions.

One remaining issue is that when (for example) a level comes with a
font.png but not a font.txt, the game will load the level-specific
font.png along with the stock font.txt that now exists, because they're
both next to each other in the virtual filesystem. That's something
that can already be seen if you have font.png+font.txt in your
Documents/VVVVVV/graphics folder and load a level with a custom
font.png. But that's easily fixed by an added check that both must be
in the same physical folder, so I'll add that as well.