Prevent font.png and font.txt from different dirs being used together
I added `FILESYSTEM_areAssetsInSameRealDir(filenameA, filenameB)` and
used it to make sure font.png and font.txt have to be in the same
folder for font.txt to be accepted. The result is that font.pngs
without font.txt (like in levels) don't inherit the font.txt from
somewhere further in the search path (like the stock font.txt, but also
a font.png+font.txt in the Documents/VVVVVV folder which was already
possible) and make the text all messed up.

This technically goes the other way as well - a stock font needing
font.txt could be forced to go without it by a level or graphics folder
supplying font.txt but not font.png. But that would be a strange and
incredibly rare setup - a sole font.txt is pointless.