Add missing limits to strings
If a limit is missing from a string, I want there to be at least a
reason for it - which can be because it's one of the practically
undefinable menu options, because the limit depends on the lengths of
other strings (like Low/Medium/High in the joystick menu), or a string
looks progressively worse the longer it gets. As a general rule: if
defining a hard limit would be misleading, then it can be exempt from
having a limit so the translator knows to pay attention to how and
where that string actually appears.

I also did some other general fixes, there was one or two strings with
an outdated limit, and there were two different versions of the
"Enter map filename to save as:" string which I merged into one, and
I added one or two explanations.