Add ALWAYS_SHOW_TRANSLATOR_MENU define
There were already two triggers for the translator menu to show up,
namely the "lang" folder being found in "desktop_version" instead of
next to data.zip, or passing -translator as a command line argument.

I thought this was good enough, since, as far as my knowledge goes,
Dicey Dungeons also requires passing -translator. But at the very
least it adds a step to the translators' instructions, and not every
translator may as easily know how to do this. Alternatively, we could
send them a version wrapped in a fake desktop_version folder, but it'd
need a warning not to take it out or rename the folder. So it would be
good to make the process smoother.

Inspired by the `#define MAKEANDPLAY` approach, it's now also possible
to define ALWAYS_SHOW_TRANSLATOR_MENU at the top of Localization.h.
That way, translators can comfortably be sent a special version of the
game without error-prone steps on how to unlock translator mode,
because it'll be enabled by default.