Remove option(...) variant for std::string
This was needed when loc::gettext still returned a std::string instead
of a const char*, to avoid adding .c_str() in a million places. This
variant has now become almost unused though, only two lines still
relied on it, so I've added the .c_str()s and removed it.